+++ src/main/java/roomescape/member/application/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.member.presentation.dto;
+package roomescape.member.application.dto;
 
 import roomescape.member.domain.Member;
