+++ src/main/java/roomescape/member/application/dto/TokenRequest.java
@@ -1,9 +1,9 @@
-package roomescape.member.presentation.dto;
+package roomescape.member.application.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
-public class SignUpRequest {
+public class TokenRequest {
 
     @Email(message = "이메일 형식이 올바르지 않습니다")
     private final String email;
@@ -11,13 +11,9 @@ public class SignUpRequest {
     @NotBlank(message = "비밀번호는 공백일 수 없습니다")
     private final String password;
 
-    @NotBlank(message = "이름은 공백일 수 없습니다")
-    private final String name;
-
-    public SignUpRequest(String email, String password, String name) {
+    public TokenRequest(String email, String password) {
         this.email = email;
         this.password = password;
-        this.name = name;
     }
 
     public String getEmail() {
@@ -26,9 +22,5 @@ public class SignUpRequest {
 
     public String getPassword() {
         return password;
-    }
-
-    public String getName() {
-        return name;
     }
 }
