+++ src/main/java/roomescape/payment/application/service/PaymentClient.java
@@ -1,7 +1,7 @@
 package roomescape.payment.application.service;
 
 import roomescape.payment.domain.Payment;
-import roomescape.payment.presentation.dto.PaymentRequest;
+import roomescape.payment.application.dto.PaymentRequest;
 
 public interface PaymentClient {
