+++ src/main/java/roomescape/reservation/application/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
-package roomescape.reservation.presentation.dto;
+package roomescape.reservation.application.dto;
 
 import java.time.LocalDate;
-import roomescape.member.presentation.dto.MemberResponse;
+import roomescape.member.application.dto.MemberResponse;
 import roomescape.reservation.domain.Reservation;
 
 public class ReservationResponse {
