+++ src/main/java/roomescape/reservation/application/dto/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.presentation.dto;
+package roomescape.reservation.application.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
