+++ src/main/java/roomescape/reservation/application/dto/UserReservationsResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.presentation.dto;
+package roomescape.reservation.application.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
