+++ src/main/java/roomescape/reservation/application/dto/WaitingRequest.java
@@ -1,9 +1,9 @@
-package roomescape.reservation.presentation.dto;
+package roomescape.reservation.application.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
-public class AdminReservationRequest {
+public class WaitingRequest {
 
     @NotNull(message = "날짜는 반드시 입력해야 합니다")
     private final LocalDate date;
@@ -14,29 +14,21 @@ public class AdminReservationRequest {
     @NotNull(message = "시작 시간은 반드시 입력해야 합니다")
     private final Long timeId;
 
-    @NotNull(message = "유저는 반드시 입력해야 합니다")
-    private final Long memberId;
-
-    public AdminReservationRequest(LocalDate date, Long themeId, Long timeId, Long memberId) {
+    public WaitingRequest(final LocalDate date, final Long themeId, final Long timeId) {
         this.date = date;
         this.themeId = themeId;
         this.timeId = timeId;
-        this.memberId = memberId;
     }
 
     public LocalDate getDate() {
         return date;
     }
 
-    public Long getTimeId() {
-        return timeId;
-    }
-
     public Long getThemeId() {
         return themeId;
     }
 
-    public Long getMemberId() {
-        return memberId;
+    public Long getTimeId() {
+        return timeId;
     }
 }
