+++ src/test/java/roomescape/member/presentation/MemberControllerTest.java
@@ -14,7 +14,7 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
 import roomescape.DatabaseCleaner;
-import roomescape.member.presentation.dto.SignUpRequest;
+import roomescape.member.application.dto.SignUpRequest;
 import roomescape.member.presentation.fixture.MemberFixture;
 
 @ActiveProfiles("test")
