+++ src/test/java/roomescape/payment/StubPaymentClient.java
@@ -2,7 +2,7 @@ package roomescape.payment;
 
 import roomescape.payment.application.service.PaymentClient;
 import roomescape.payment.domain.Payment;
-import roomescape.payment.presentation.dto.PaymentRequest;
+import roomescape.payment.application.dto.PaymentRequest;
 
 public class StubPaymentClient implements PaymentClient {
     @Override
