+++ src/test/java/roomescape/reservation/presentation/AdminReservationControllerTest.java
@@ -17,7 +17,7 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.DatabaseCleaner;
 import roomescape.TestConfig;
 import roomescape.member.presentation.fixture.MemberFixture;
-import roomescape.reservation.presentation.dto.AdminReservationRequest;
+import roomescape.reservation.application.dto.AdminReservationRequest;
 import roomescape.reservation.presentation.fixture.ReservationFixture;
 
 @ActiveProfiles("test")
