+++ src/test/java/roomescape/reservation/presentation/ReservationControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.DatabaseCleaner;
 import roomescape.TestConfig;
 import roomescape.member.presentation.fixture.MemberFixture;
-import roomescape.reservation.presentation.dto.ReservationRequest;
+import roomescape.reservation.application.dto.ReservationRequest;
 import roomescape.reservation.presentation.fixture.ReservationFixture;
 
 @ActiveProfiles("test")
