+++ src/test/java/roomescape/reservation/presentation/ReservationTimeControllerTest.java
@@ -20,7 +20,7 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.DatabaseCleaner;
 import roomescape.TestConfig;
 import roomescape.member.presentation.fixture.MemberFixture;
-import roomescape.reservation.presentation.dto.ReservationTimeRequest;
+import roomescape.reservation.application.dto.ReservationTimeRequest;
 import roomescape.reservation.presentation.fixture.ReservationFixture;
 
 @ActiveProfiles("test")
