+++ src/test/java/roomescape/reservation/presentation/ThemeControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.DatabaseCleaner;
 import roomescape.TestConfig;
 import roomescape.member.presentation.fixture.MemberFixture;
-import roomescape.reservation.presentation.dto.ThemeRequest;
+import roomescape.reservation.application.dto.ThemeRequest;
 import roomescape.reservation.presentation.fixture.ReservationFixture;
 
 @ActiveProfiles("test")
