+++ src/test/java/roomescape/reservation/presentation/fixture/ReservationFixture.java
@@ -5,11 +5,11 @@ import io.restassured.http.ContentType;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Map;
-import roomescape.reservation.presentation.dto.ReservationRequest;
-import roomescape.reservation.presentation.dto.AdminReservationRequest;
-import roomescape.reservation.presentation.dto.ReservationTimeRequest;
-import roomescape.reservation.presentation.dto.ThemeRequest;
-import roomescape.reservation.presentation.dto.WaitingRequest;
+import roomescape.reservation.application.dto.ReservationRequest;
+import roomescape.reservation.application.dto.AdminReservationRequest;
+import roomescape.reservation.application.dto.ReservationTimeRequest;
+import roomescape.reservation.application.dto.ThemeRequest;
+import roomescape.reservation.application.dto.WaitingRequest;
 
 public class ReservationFixture {
